refactor: Move out plan_modifer internal functions
